feat: add subscribe retry logic, clean up code
