+++ examples/settings.gradle.kts
@@ -14,13 +14,13 @@ rootProject.name = "java-gradle"
  * copy of the SDK code. This is useful if you are making changes to the SDK and want to
  * test them in the examples before committing them.
  */
-includeBuild("..") {
-    dependencySubstitution {
-        substitute(module("software.momento.java:sdk")).using(project(":momento-sdk"))
-    }
-}
+//includeBuild("..") {
+//    dependencySubstitution {
+//        substitute(module("software.momento.java:sdk")).using(project(":momento-sdk"))
+//    }
+//}
 
 include("cache")
 include("cache-with-aws")
 include("lambda:docker")
-
+include("topic")
