+++ momento-sdk/src/main/java/momento/sdk/ScsTopicGrpcStubsManager.java
@@ -8,6 +8,7 @@ import java.io.Closeable;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 import javax.annotation.Nonnull;
 import momento.sdk.auth.CredentialProvider;
 import momento.sdk.config.TopicConfiguration;
@@ -21,16 +22,16 @@ import momento.sdk.config.TopicConfiguration;
  */
 final class ScsTopicGrpcStubsManager implements Closeable {
 
-  private final Duration deadline;
-
   private final ManagedChannel channel;
   private final PubsubGrpc.PubsubStub stub;
+
+  private final TopicConfiguration configuration;
 
   ScsTopicGrpcStubsManager(
       @Nonnull CredentialProvider credentialProvider, @Nonnull TopicConfiguration configuration) {
     this.channel = setupConnection(credentialProvider);
     this.stub = PubsubGrpc.newStub(channel);
-    this.deadline = configuration.getTransportStrategy().getGrpcConfiguration().getDeadline();
+    this.configuration = configuration;
   }
 
   private static ManagedChannel setupConnection(CredentialProvider credentialProvider) {
@@ -38,6 +39,13 @@ final class ScsTopicGrpcStubsManager implements Closeable {
         NettyChannelBuilder.forAddress(credentialProvider.getCacheEndpoint(), 443);
     channelBuilder.useTransportSecurity();
     channelBuilder.disableRetry();
+
+    System.out.println("\n\n\n\nSETTING KEEPALIVES\n\n\n\n");
+
+    channelBuilder.keepAliveTime(10, TimeUnit.SECONDS);
+    channelBuilder.keepAliveTimeout(5, TimeUnit.SECONDS);
+    channelBuilder.keepAliveWithoutCalls(true);
+
     final List<ClientInterceptor> clientInterceptors = new ArrayList<>();
     clientInterceptors.add(new UserHeaderInterceptor(credentialProvider.getAuthToken()));
     channelBuilder.intercept(clientInterceptors);
@@ -57,8 +65,10 @@ final class ScsTopicGrpcStubsManager implements Closeable {
   PubsubGrpc.PubsubStub getStub() {
     return stub;
   }
-  //    .withDeadlineAfter(deadline.getSeconds(), TimeUnit.SECONDS);
-  //  }
+
+  TopicConfiguration getConfiguration() {
+    return configuration;
+  }
 
   @Override
   public void close() {
